Moved route handler out into separate controller files for each entity. Also, added .babelrc file and updated controllers to be classes, and require() and module.export statements are now written as import and export statements.
